Adding Tag entity

+++ src/main/java/com/luxcampus/blog/repository/PostRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
 
     List<Post> findByTitle(String title);
